fix: ct_repo_lint compliance for Refs #3093 slice 2
- Switch `world_market_sellable_quantity_test.dart` to the
  `package:colonizethis_test/test.dart` entrypoint required by
  `[repo.test_imports]` per `SPEC/program/test-logging.md`.
- Split the Deal Book widgets out of `trade_screen.dart` into a
  sibling `trade_screen_deal_book.dart` part fragment so the host
  file drops back under the 1000 non-comment line cap enforced by
  `[repo.dart_file_non_comment_line_size]`
  (`SPEC/program/dart-file-non-comment-line-size.md`). All
  `_DealBook*` classes stay library-private via `part of` and
  continue to reference the same `TradeScreen.dealBook*` constants
  unchanged.

Refs #3093
