+++ packages/colonizethis_logic/test/world_market_sellable_quantity_test.dart
@@ -6,7 +6,7 @@
 
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:test/test.dart';
+import 'package:colonizethis_test/test.dart';
 
 const String _humanPlayerId = 'gp_h';
